Adding a new layer type: Dummy.
This layer type can be used as a placeholder for other module, so there's no need to provide placeholder layer by other module.
Exemple, with Geofield:
With hook_openlayers_map_preprocess, we could replace the dummy layer with the feature layer.
Of course, the formatter settings will have to be updated.